Fix type. Add print samples and pdf. Add config to control test data

+++ utils/utils.py
@@ -221,8 +221,8 @@ def map_to_hemisphere(s: np.ndarray, n: np.ndarray, e=1.):
 def get_test_samples(points: np.ndarray):
 
     norm_pts = points[:, 3:]
-    lights = np.zeros(norm_pts.shape)
-    pdfs = np.zeros(norm_pts.shape[0])
+    lights = np.zeros(norm_pts.shape, dtype=np.float32)
+    pdfs = np.zeros(norm_pts.shape[0], dtype=np.float32)
 
     for i in range(norm_pts.shape[0]):
         s = np.random.uniform(0., 1., 2)
@@ -230,5 +230,4 @@ def get_test_samples(points: np.ndarray):
         light = map_to_hemisphere(s, norm)
         lights[i] = light
         pdfs[i] = np.dot(norm, light) / np.linalg.norm(norm) / np.linalg.norm(light) / np.pi
-
     return [torch.from_numpy(lights).to('cpu'), torch.from_numpy(pdfs).to('cpu')]
